Update README.md
On the migration to use bigint: added that this can take hours and will increase the database size substantially. (Example with a real database running on a linode: 40GB => 200GB (without analytics) and over 5 hours). People (like me) might think that the startup is stuck, since there is not message in the tomcat log that this is happening.